Explore using external scanner for parsing <template> content
It allows greater control and allows us to expose more of the template.
External scanner seems to be the way if we want to expose more than just
the `<template>` token.

This obviously makes it more tedious to update the upstream TypeScript
grammar. Especially the `GlimmerTokenType` which must continue the
sequence of the parent's TokenType.